task: remove old ir versions
