+++ libs/shared/schemas/innovation-record/translation.helper.ts
@@ -1,6 +1,3 @@
-import Joi from 'joi';
-import { DocumentValidationSchema202304Map } from './document.schema';
-
 // This will be replaced in the future this is only here in the meanwhile while the IR gets refactored
 const categoriesTranslation = {
   MEDICAL_DEVICE: 'Medical device',
@@ -476,11 +473,3 @@ export const translateValue = (fields: string[], value: string): string => {
   return typeof translated === 'string' ? translated : value;
 };
 
-// This is a naive implementation of the cleanup, not doing anything on the validation errors for now
-export const cleanup = (source: Record<string, any>): any => {
-  const result = Joi.object(DocumentValidationSchema202304Map).validate(source, {
-    abortEarly: false,
-    stripUnknown: true
-  });
-  return result.value;
-};
